+++ spring_jdbc/src/main/java/com/packt/footballpg/records/Team.java
@@ -1,4 +1,4 @@
 package com.packt.footballpg.records;
 
-public record Team(Integer id, String name) {
+public record Team(Long id, String name) {
 }
